+++ src/main/java/school/faang/sweetHogwars/spellAction.java
@@ -1,0 +1,6 @@
+package school.faang.sweetHogwars;
+
+@FunctionalInterface
+interface spellAction {
+    String perform(String spellName);
+}
